requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
google-generativeai>=0.7.0
